+++ CHANGES.md
@@ -5,13 +5,14 @@ chronological order. Releases follow [semantic versioning](https://semver.org/) 
 releases are available on [PyPI](https://pypi.org/project/pytask-parallel) and
 [Anaconda.org](https://anaconda.org/conda-forge/pytask-parallel).
 
-## 0.4.0 - 2023-xx-xx
+## 0.4.0 - 2023-10-07
 
 - {pull}`62` deprecates Python 3.7.
 - {pull}`64` aligns pytask-parallel with pytask v0.4.0rc2.
 - {pull}`66` deactivates parallelization for dry-runs.
 - {pull}`67` fixes parallelization with partialed task functions.
-- {pull}`68` updates dependencies and syntaxes.
+- {pull}`68` raises more informative error message when `breakpoint()` was uses when
+  parallelizing with processes or loky.
 
 ## 0.3.1 - 2023-05-27